Completed

+++ frontend/src/components/WorkoutDetails/index.js
@@ -2,13 +2,20 @@ import { type } from "../../context/constants";
 import { useWorkoutContext } from "../../hooks/WorkoutHook";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 import "./index.css";
+import { useAuthContext } from "../../hooks/AuthHook";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) return;
+
     const res = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        'Authorization': `Bearer ${user.token}`,
+      },
     });
     const json = await res.json();
